Bump `syn` dep to the latest version (#1)
Your `Cargo.lock` is currently locking an old version of the `syn`
crate. In future versions of Rust (in particular, once
https://github.com/rust-lang/rust/pull/77153) is merged), this may cause
your crate to stop compiling.

This PR bumps `syn` to the latest version in your `Cargo.lock`. This
ensures that your crate can compiler with both older and newer versions
of Rust.